refactor: move user repository code
