+++ src/main/java/com/spring/practice/rest/repository/user/impl/JdbcTemplateUserRepository.java
@@ -1,4 +1,4 @@
-package com.spring.practice.rest.repository.impl;
+package com.spring.practice.rest.repository.user.impl;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import com.spring.practice.rest.domain.user.User;
-import com.spring.practice.rest.repository.UserRepository;
+import com.spring.practice.rest.repository.user.UserRepository;
 
 @Repository("JdbcTemplateUserRepository")
 public class JdbcTemplateUserRepository implements UserRepository {
